refactor: Move evaluation to interpreter module
