+++ src/core/constraints/mod.rs
@@ -2,4 +2,3 @@ pub mod constraint;
 pub mod constrainted_type;
 pub mod sampler;
 pub mod tests;
-mod util;
