+++ src/core/mod.rs
@@ -1,2 +1,5 @@
 pub mod ast;
+pub mod constraints;
+pub mod model;
 pub mod parser;
+pub mod utils;
