+++ src/core/model/mod.rs
@@ -1,0 +1,2 @@
+pub mod enumeration;
+pub mod template;
