+++ src/core/utils/mod.rs
@@ -1,0 +1,1 @@
+pub mod extract_object;
