+++ src/lib.rs
@@ -1,8 +1,4 @@
-pub mod constraints;
-pub mod enumeration;
-pub mod evaluator;
-pub mod generation;
-
 pub mod core;
-pub mod template;
+pub mod generation;
+pub mod interpreter;
 pub mod tests;
