+++ src/tests/lexer_tests.rs
@@ -1,4 +1,3 @@
-
 use std::vec;
 
 use crate::core::parser::{
