+++ src/tests/mod.rs
@@ -3,3 +3,6 @@ pub mod lexer_tests;
 
 #[cfg(test)]
 pub mod parser_tests;
+
+#[cfg(test)]
+pub mod interpreter_tests;
